requirements.txt
opencv-python
pandas
convertbng
pynmeagps
. -e
